Move test_sampling fixtures to conftest file
